Improve Math.random randomness.
Largest possible MLCG that only uses integer arithmetic with doubles.

Parameters from Pierre l'Ecuyers paper "Tables of linear congruential
generators of different sizes and good lattice structure".

+++ jsi.h
@@ -239,7 +239,7 @@ struct js_State
 	js_Object *TypeError_prototype;
 	js_Object *URIError_prototype;
 
-	unsigned int seed; /* Math.random seed */
+	double seed; /* Math.random seed */
 
 	int nextref; /* for js_ref use */
 	js_Object *R; /* registry of hidden values */
